Pin demo to specific trip tracker aar
While a specific version from maven (e.g. `implementation
'com.joinroot:root-android-trip-tracker:4.9.0'`) is preferable to a
hard-coded file, doing this allows us to pull and run the demo without
needing to worry about fixing breaking changes or hunting down the last
specific version this app worked with.

It's worth noting that this still doesn't make it obvious WHAT the last
version of trip tracker used in this demo is (for reference, the aar in
this commit is release 4.9.0), but that's an existing problem.