Use Room database for storing log messages
Note that allowMainThreadQueries on line 26 is NOT recommended. The more
proper/correct way of doing this is to have the DAO implement LiveData
streams and handle those in async queries instead of on the main thread.
This felt like overkill for our needs here, but is worth calling out if
we intend to expand on this in the future